chore: table的一些修改

+++ internal/service/user.go
@@ -1,10 +1,10 @@
 package service
 
 type User struct {
-	Id            int64  `json:"id,omitempty"`
+	Id            uint64 `json:"id,omitempty"`
 	Name          string `json:"name,omitempty"`
-	FollowCount   int64  `json:"follow_count,omitempty"`
-	FollowerCount int64  `json:"follower_count,omitempty"`
+	FollowCount   uint32 `json:"follow_count,omitempty"`
+	FollowerCount uint32 `json:"follower_count,omitempty"`
 	IsFollow      bool   `json:"is_follow,omitempty"`
 }
 
@@ -15,7 +15,7 @@ type UserListResponse struct {
 
 type UserLoginResponse struct {
 	Response
-	UserId int64  `json:"user_id,omitempty"`
+	UserId uint64 `json:"user_id,omitempty"`
 	Token  string `json:"token"`
 }
 
@@ -32,18 +32,18 @@ func (s *Service) Login(username string, password string) UserLoginResponse {
 	return UserLoginResponse{}
 }
 
-func (s *Service) Follow(userId int64, to_userId int64) Response {
+func (s *Service) Follow(userId uint64, to_userId uint64) Response {
 	return Response{}
 }
 
-func (s *Service) GetFollowList(userId int64) UserListResponse {
+func (s *Service) GetFollowList(userId uint64) UserListResponse {
 	return UserListResponse{}
 }
 
-func (s *Service) GetFollowerList(userId int64) UserListResponse {
+func (s *Service) GetFollowerList(userId uint64) UserListResponse {
 	return UserListResponse{}
 }
 
-func (s *Service) GetUserInfo(userId int64) UserResponse {
+func (s *Service) GetUserInfo(userId uint64) UserResponse {
 	return UserResponse{}
 }
