+++ internal/service/video.go
@@ -3,12 +3,12 @@ package service
 import "time"
 
 type Video struct {
-	Id            int64  `json:"id,omitempty"`
+	Id            uint64 `json:"id,omitempty"`
 	Author        User   `json:"author"`
 	PlayUrl       string `json:"play_url,omitempty"`
 	CoverUrl      string `json:"cover_url,omitempty"`
-	FavoriteCount int64  `json:"favorite_count,omitempty"`
-	CommentCount  int64  `json:"comment_count,omitempty"`
+	FavoriteCount uint32 `json:"favorite_count,omitempty"`
+	CommentCount  uint32 `json:"comment_count,omitempty"`
 	IsFavorite    bool   `json:"is_favorite,omitempty"`
 }
 
@@ -23,7 +23,7 @@ type VideoListResponse struct {
 	VideoList []Video `json:"video_list"`
 }
 
-func (s *Service) Feed(userId int64, lasttime time.Time) FeedResponse {
+func (s *Service) Feed(userId uint64, lasttime time.Time) FeedResponse {
 	return FeedResponse{
 		Response:  Response{StatusCode: 0},
 		VideoList: []Video{},
@@ -31,22 +31,22 @@ func (s *Service) Feed(userId int64, lasttime time.Time) FeedResponse {
 	}
 }
 
-func (s *Service) PublishVideo(userId int64, video Video) Response {
+func (s *Service) PublishVideo(userId uint64, video Video) Response {
 	return Response{StatusCode: 0}
 }
 
-func (s *Service) LikeDisliakeVideo(userId int64, videoId int64, like bool) Response {
+func (s *Service) LikeDisliakeVideo(userId uint64, videoId uint64, like bool) Response {
 	return Response{StatusCode: 0}
 }
 
-func (s *Service) GetVideoList(userId int64) VideoListResponse {
+func (s *Service) GetVideoList(userId uint64) VideoListResponse {
 	return VideoListResponse{
 		Response:  Response{StatusCode: 0},
 		VideoList: []Video{},
 	}
 }
 
-func (s *Service) GetLikeList(userId int64) VideoListResponse {
+func (s *Service) GetLikeList(userId uint64) VideoListResponse {
 	return VideoListResponse{
 		Response:  Response{StatusCode: 0},
 		VideoList: []Video{},
